+++ changelog.md
@@ -7,6 +7,7 @@
 * `FIX` missed syntax error `f() = 1`
 * `FIX` missed global `bit` in `LuaJIT`
 * `FIX` completion: may insert error text when continuous inputing
+* `FIX` completion: may insert error text after resolve
 * `FIX` [#349](https://github.com/sumneko/lua-language-server/issues/349)
 
 ## 1.15.1